backend/requirements.txt
# FastAPI Core
fastapi==0.135.2
starlette==0.52.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
pydantic-settings==2.13.1
jinja2==3.1.5
python-multipart==0.0.20
aiofiles==24.1.0

# Database — SQLAlchemy 2.0 / SQLModel / Alembic on PostgreSQL
sqlalchemy==2.0.48
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.11

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# HTTP & API
httpx>=0.24.0
requests==2.32.3
aiohttp==3.11.11

# Task Queue
celery==5.6.2
redis==7.3.0

# Integrations
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
google-api-python-client==2.193.0

# OCR & AI
openai==2.29.0
anthropic==0.86.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==12.1.1

# Data Processing
pandas==3.0.1
numpy==2.4.3

# Utilities
python-dateutil==2.9.0
pytz==2025.2
pydantic-extra-types==2.10.0

# Development
pytest==8.4.1
pytest-asyncio==0.25.2
pytest-cov==6.1.1
black==25.1.0
ruff==0.11.10
email-validator>=2.0.0
